# DBA.Web/src/api/generated-client.ts
// AUTO-GENERATED — do not edit manually. Run `npm run api:sync` to regenerate.
import type { components } from './schema';

export type DeviceBookingDto = components['schemas']['DeviceBookingDto'];
export type CreateDeviceBookingCommand = components['schemas']['CreateDeviceBookingCommand'];
export type ApproveDeviceBookingCommand = components['schemas']['ApproveDeviceBookingCommand'];
export type RejectDeviceBookingCommand = components['schemas']['RejectDeviceBookingCommand'];